Remove duplicate getAnimation function and improve namespacing
